Refactor the entire logic of the calculator
Prior to this commit, my inexperience lead to very unwieldy code
that depended heavily on click functions as opposed to streamlining
the functions to accept the button content. This also allows for
keyboard support. In addition, the running solution has been changed
to a running display, displaying the equation of the first operand,
which is effectively the running solution, the operator, and the second operand.